// frontend/src/service/api/models/CommentRes.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CommentRes = {
    id: number;
    content_id: number;
    creator_id: number;
    parent_id: number;
    text: string;
    replies: Array<CommentRes>;
    created_at: string;
};